Moved fashion df function to dataset
