+++ few_shot/experiments/fashion/data_augmentation_more_patience_save_latest.py
@@ -4,7 +4,8 @@ import numpy as np
 import pandas as pd
 import tensorflow as tf
 
-from few_shot.experiments.fashion import evaluate_fashion_few_shot, fashion_dfs
+from few_shot.experiments.fashion import evaluate_fashion_few_shot
+from few_shot.dataset.fashion import fashion_dfs
 from few_shot.dataset.image_pipeline import augmented_img_pipeline_fn
 
 if __name__ == '__main__':
