+++ few_shot/experiments/fashion/hyperparameter_search.py
@@ -5,7 +5,8 @@ import pandas as pd
 import skopt
 import tensorflow as tf
 
-from few_shot.experiments.fashion import evaluate_fashion_few_shot, fashion_dfs
+from few_shot.experiments.fashion import evaluate_fashion_few_shot
+from few_shot.dataset.fashion import fashion_dfs
 from few_shot.dataset.image_pipeline import augmented_img_pipeline_fn
 
 MAX_SHOTS = 5
